requirements.txt
Django>=1.4.5
South>=0.8.4
django-modeltranslation>=0.7.3
psycopg2>=2.5.2
